Extract the direction dropdown to its own shared component for asset and dag graphs.
