+++ airflow-core/src/airflow/ui/src/layouts/Details/PanelButtons.tsx
@@ -42,11 +42,10 @@ import { useParams } from "react-router-dom";
 import { useLocalStorage } from "usehooks-ts";
 
 import { DagVersionSelect } from "src/components/DagVersionSelect";
-import { directionOptions, type Direction } from "src/components/Graph/useGraphLayout";
 import { Tooltip } from "src/components/ui";
 import { type ButtonGroupOption, ButtonGroupToggle } from "src/components/ui/ButtonGroupToggle";
 import type { DagView } from "src/constants/dagView";
-import { dependenciesKey, directionKey } from "src/constants/localStorage";
+import { dependenciesKey } from "src/constants/localStorage";
 import type { VersionIndicatorOptions } from "src/constants/showVersionIndicatorOptions";
 import { useContainerWidth } from "src/utils/useContainerWidth";
 
@@ -56,6 +55,7 @@ import { GridFilters } from "./GridFilters";
 import { TaskStreamFilter } from "./TaskStreamFilter";
 import { ToggleGroups } from "./ToggleGroups";
 import { VersionIndicatorSelect } from "./VersionIndicatorSelect";
+import { DirectionDropdown } from "src/components/Graph/DirectionDropdown";
 
 type Props = {
   readonly dagView: DagView;
@@ -117,7 +117,6 @@ export const PanelButtons = ({
     dependenciesKey(dagId),
     "tasks",
   );
-  const [direction, setDirection] = useLocalStorage<Direction>(directionKey(dagId), "RIGHT");
   const containerRef = useRef<HTMLDivElement>(null);
   const containerWidth = useContainerWidth(containerRef);
   const handleLimitChange = (event: SelectValueChangeDetails<{ label: string; value: Array<string> }>) => {
@@ -144,14 +143,6 @@ export const PanelButtons = ({
       removeDependencies();
     } else {
       setDependencies(event.value[0]);
-    }
-  };
-
-  const handleDirectionUpdate = (
-    event: SelectValueChangeDetails<{ label: string; value: Array<string> }>,
-  ) => {
-    if (event.value[0] !== undefined) {
-      setDirection(event.value[0] as Direction);
     }
   };
 
@@ -277,34 +268,7 @@ export const PanelButtons = ({
                           </Select.Positioner>
                         </Select.Root>
 
-                        <Select.Root
-                          // @ts-expect-error The expected option type is incorrect
-                          collection={directionOptions(translate)}
-                          onValueChange={handleDirectionUpdate}
-                          size="sm"
-                          value={[direction]}
-                        >
-                          <Select.Label fontSize="xs">
-                            {translate("dag:panel.graphDirection.label")}
-                          </Select.Label>
-                          <Select.Control>
-                            <Select.Trigger>
-                              <Select.ValueText />
-                            </Select.Trigger>
-                            <Select.IndicatorGroup>
-                              <Select.Indicator />
-                            </Select.IndicatorGroup>
-                          </Select.Control>
-                          <Select.Positioner>
-                            <Select.Content>
-                              {directionOptions(translate).items.map((option) => (
-                                <Select.Item item={option} key={option.value}>
-                                  {option.label}
-                                </Select.Item>
-                              ))}
-                            </Select.Content>
-                          </Select.Positioner>
-                        </Select.Root>
+                        <DirectionDropdown graphId={dagId} />
                       </>
                     ) : (
                       <>
